README: Remove status badge
I don't think this badge has shown the actual status of the current
workflows for a long time. The real status can easily be seen by
clicking on the checkmark or cross icon on the front page. It's also
supposed to show the status of the master branch, not develop (default).